ingestion/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finnhub-pipeline"
version = "0.1.0"
description = "Finnhub ingestion jobs for market analytics"
requires-python = ">=3.10"
dependencies = [
  "finnhub-python>=2.4.24",
  "yfinance>=0.2.52",
  "pandas>=2.2.0",
  "pyarrow>=18.0.0",
  "google-cloud-storage>=2.19.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
